[김다은] update hompage, mycompanycomparison
