+++ client/src/components/Dropdown.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useRef, useEffect } from "react";
 import styled, { css, keyframes } from "styled-components";
 import arrowDown from "../assets/images/dropdownarrow/arrow.svg";
+import { gray_200 } from "../styles/colors";
 
 const SORT_OPTIONS = [
   "View My Startup 투자 금액 높은순",
@@ -64,7 +65,7 @@ const DropdownButton = styled.button`
 
   background-color: #000;
   color: #fff;
-  border: 1px solid #fff;
+  border: 1px solid ${gray_200};
   border-radius: 14px;
   width: ${(props) => (props.$size === "big" ? "280px" : "240px")};
   height: ${(props) => (props.$size === "big" ? "44px" : "39px")};
@@ -137,8 +138,8 @@ const DropdownItem = styled.button`
   }
 
   &:hover {
-    background-color: #fff;
-    color: #000;
+    background-color: #212121;
+    color: #fff;
   }
 `;
