+++ client/src/pages/MainFullList/Homepage.module.css
@@ -1,5 +1,5 @@
 body {
-  background-color: #000;
+  background-color: #111;
 }
 
 .container {
