--- client/src/pages/MainFullList/index.jsx
@@ -1,86 +1,0 @@
-import styles from "./Homepage.module.css";
-import TableHeader from "../../components/TableHeader";
-import Search from "../../components/Search";
-import BtnPagination from "../../components/BtnPagination";
-import React, { useEffect, useState } from "react";
-import SortDropdown from "../../components/Dropdown";
-
-function HomePage() {
-  const columns = [
-    { label: "순위", name: "ranking", flex: 0.5 },
-    { label: "기업명", name: "name", flex: 1.5 },
-    { label: "기업 소개", name: "description", flex: 4 },
-    { label: "카테고리", name: "category", flex: 2 },
-    { label: "누적 투자 금액", name: "investmentAmount", flex: 1 },
-    { label: "매출액", name: "revenue", flex: 2 },
-    { label: "고용 인원", name: "employees", flex: 1.5 },
-  ];
-
-  const [mediaSize, setMediaSize] = useState("");
-
-  function updateMediaSize() {
-    const { innerWidth: width } = window;
-    if (width > 744) {
-      setMediaSize("medium");
-    } else {
-      setMediaSize("small");
-    }
-  }
-
-  useEffect(() => {
-    updateMediaSize();
-
-    window.addEventListener("resize", updateMediaSize);
-
-    return () => {
-      window.removeEventListener("resize", updateMediaSize);
-    };
-  }, []);
-
-  return (
-    <body>
-      <div className={styles.container}>
-        {/* navigation bar는 추후 수정할 예정 */}
-
-        <div className={styles.content}>
-          <div className={styles.head}>
-            <div className={styles.title}>전체 스타트업 목록</div>
-
-            <div className={styles.controls}>
-              <Search size={mediaSize} state={"none"} />
-              <div>
-                <SortDropdown size={"medium"} />
-              </div>
-            </div>
-          </div>
-
-          <table className={styles.table}>
-            <thead>
-              <TableHeader columns={columns} />
-            </thead>
-            <tbody>
-              {/* 데이터 자리 */}
-              <tr>
-                <td
-                  colSpan={columns.length}
-                  style={{
-                    textAlign: "center",
-                    padding: "40px 0",
-                    color: "#888",
-                  }}
-                >
-                  표시할 스타트업 데이터가 없습니다.
-                </td>
-              </tr>
-            </tbody>
-          </table>
-          <div className={styles.pagination}>
-            <BtnPagination size={"big"} />
-          </div>
-        </div>
-      </div>
-    </body>
-  );
-}
-
-export default HomePage;
